changed `SovereignStateRankedValue` value type from `Double` to `Float`, and... - removed `value_type` from `SovereignStateAgricultureValue` and `SovereignStateRankingInfoValue`

+++ Sources/ParadigmUtilities/jsons/countries/information/values/SovereignStateRankedValue.swift
@@ -10,7 +10,7 @@ import Foundation
 public protocol SovereignStateRankedValue : SovereignStateIdentifiableValue {
     var world_rank:Int { get }
     var world_rank_max:Int { get }
-    var value:Double { get }
+    var value:Float { get }
     var suffix:String? { get }
     var is_estimate:Bool { get }
 }
